condensed login contexts, changed longest streak label, updated imports for progress-related code
