+++ src/homepage.js
@@ -16,13 +16,9 @@ import { dailyIncrement } from "./progress.js";
 import { useDailyLogins } from './progress_files/dailyLoginsContext.js';
 import { useConsecutiveLogins } from './progress_files/consecutiveLoginsContext.js';
 import { useLongestStreak } from "./progress_files/longestStreakContext.js";
-import { useSundayLogin } from './progress_files/sundayLoginContext.js';
-import { useMondayLogin } from './progress_files/mondayLoginContext.js';
-import { useTuesdayLogin } from './progress_files/tuesdayLoginContext.js';
-import { useWednesdayLogin } from './progress_files/wednesdayLoginContext.js';
-import { useThursdayLogin } from './progress_files/thursdayLoginContext.js';
-import { useFridayLogin } from './progress_files/fridayLoginContext.js';
-import { useSaturdayLogin } from './progress_files/saturdayLoginContext.js';
+import { useSundayLogin, useMondayLogin, useTuesdayLogin, 
+         useWednesdayLogin, useThursdayLogin, 
+         useFridayLogin, useSaturdayLogin } from './progress_files/weeklyLoginContext.js';
 
 //Main home page 
 export const HomePage = ({navigation}) =>{
